tools/omop-tools/requirements.txt
# Run: pip3 install -r requirements.txt to install listed dependencies
# For now, versionless

argparse
pandas
google
google-cloud
google-cloud-bigquery
google-cloud-storage
google-api-python-client
sqlalchemy
python-dotenv
psycopg2-binary
